+++ src/main/c/hdr_histogram.h
@@ -304,7 +304,7 @@ bool hdr_linear_iter_next(struct hdr_linear_iter* linear);
 struct hdr_log_iter
 {
     struct hdr_iter iter;
-    int value_units_first_bucket;
+    int64_t value_units_first_bucket;
     double log_base;
     int64_t count_added_in_this_iteration_step;
     int64_t next_value_reporting_level;
@@ -322,7 +322,7 @@ struct hdr_log_iter
 void hdr_log_iter_init(
     struct hdr_log_iter* logarithmic,
     struct hdr_histogram* h,
-    int value_units_first_bucket,
+    int64_t value_units_first_bucket,
     double log_base);
 
 /**
